introducing common/SearchBox component
